--- a/iree/compiler/Codegen/LLVMCPU/LLVMCPUTileAndVectorizeLinalgTensorOps.cpp
+++ b/iree/compiler/Codegen/LLVMCPU/LLVMCPUTileAndVectorizeLinalgTensorOps.cpp
@@ -80,9 +80,8 @@ void LLVMCPUTileAndVectorizePass::runOnOperation() {
         linalg::LinalgTilingOptions().setTileSizeComputationFunction(
             [](OpBuilder &builder,
                Operation *operation) -> SmallVector<Value, 4> {
-              return getTileSizes(
-                  builder, operation,
-                  static_cast<unsigned>(TilingLevel::Level1Tiles));
+              return getTileSizes(builder, operation,
+                                  static_cast<unsigned>(TilingLevel::L1Tiles));
             }),
         linalg::LinalgTransformationFilter(
             Identifier::get(getWorkgroupMarker(), context),
@@ -117,7 +116,7 @@ void LLVMCPUTileAndVectorizePass::runOnOperation() {
                Operation *operation) -> SmallVector<Value, 4> {
               return getTileSizes(
                   builder, operation,
-                  static_cast<unsigned>(TilingLevel::Level2Tiles));
+                  static_cast<unsigned>(TilingLevel::VectorTiles));
             }),
         linalg::LinalgTransformationFilter(
             Identifier::get(getWorkgroupL1TileMarker(), context),