The current configurations on CPU backends (and all backends) rely on
looking at operations within the dispatch region to pick a
default. This makes it hard to set a good default to ensure good tile
+ distribute of operations. For every op a configuration method is
needed.

Instead one can just look at the loops and get an idea of the problem
size. These loops are created by tile + distribute at the flow level
using parametric tile sizes. The bounds of the loop are already
determined from the problem sizes. Ideally we would have a loop
representation that does not immediately materialize the distributed
loop, but rather maintains the loop in undistributed form, with
additional operands that specify the id/count/size to use when
materializing the distributed loop (defering the materialization to a
later convenient point).

In absense of this, in this PR, the tiled loop as generated today is
recognized by "lifting" it back to the untiled version based on the
know distribution pattern used at the flow level. From this the
original problem size is known as well and used for default
distribution without having to look at the code within the dispatch
region.