@MaheshRavishankar 2021-09-23
> > Need to fix some lit tests, but some preliminary x86 performance numbers
> > Before
> > ```
> > --------------------------------------------------------------------------------
> > MobileNetV3                                    Time             CPU   Iterations
> > --------------------------------------------------------------------------------
> > BM_call/process_time/real_time_mean         49.7 ms          168 ms            5
> > BM_call/process_time/real_time_median       50.0 ms          170 ms            5
> > BM_call/process_time/real_time_stddev      0.877 ms         3.58 ms            5
> > --------------------------------------------------------------------------------
> > SSD_Mobilenet_v2                               Time             CPU   Iterations
> > --------------------------------------------------------------------------------
> > BM_main/process_time/real_time_mean          625 ms         1656 ms            5
> > BM_main/process_time/real_time_median        628 ms         1654 ms            5
> > BM_main/process_time/real_time_stddev       8.13 ms         27.7 ms            5
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > After
> > ```
> > --------------------------------------------------------------------------------
> > MobileNetV3                                    Time             CPU   Iterations
> > --------------------------------------------------------------------------------
> > BM_call/process_time/real_time_mean         13.7 ms         75.7 ms            5
> > BM_call/process_time/real_time_median       13.9 ms         77.1 ms            5
> > BM_call/process_time/real_time_stddev      0.309 ms         2.24 ms            5
> > --------------------------------------------------------------------------------
> > SSD_Mobilenet_v2                               Time             CPU   Iterations
> > --------------------------------------------------------------------------------
> > BM_main/process_time/real_time_mean          137 ms          895 ms            5
> > BM_main/process_time/real_time_median        137 ms          899 ms            5
> > BM_main/process_time/real_time_stddev       1.06 ms         7.56 ms            5
> > ```
> 
> Awesome! how many cores ?

This is with 8 cores.

SSD Before 

![Before_ssd](https://user-images.githubusercontent.com/1663364/134595834-2ed3589a-3069-45fa-ad0d-47306d4c1d44.png)

SSD After

![After_ssd](https://user-images.githubusercontent.com/1663364/134595855-f4071ac3-4d89-4196-a6a4-adbbdd68b516.png)